Separated the code chunks for importing the counts and sample data.
Changed the expression criteria for inclusion in the analysis from >= 1CMP in >= 30 samples to >=1CMP in >= 7 samples as the higher requirement is more about transforming the distribution for use in DE analysis. I think it's okay to use the less stringent approach when looking only at the expression profiles.

Added annotation to the expression heatmap

Deleted the import of the mirbase db as not required in this analysis.

Everything runs
Everything plots
Everything knits

* still need to sort out the heat maps which won't fit on one page.